requirements.txt
# Project dependencies for the Nyaya application.
boto3
flask
pymupdf
python-dotenv
